Renamed querie.js to index.js and changed require in apps.js. Also fixed typo
